pyramid main test
- move TestMain to a main_test.go
- remove 3rd party package for sync counter
- delete temporary directory on test code
- ensure parent dir exists without ceilingDir

+++ pyramid/directory.go
@@ -27,23 +27,19 @@ func (d *directory) deleteDirRecIfEmpty(dir string) error {
 
 	for dir != d.ceilingDir {
 		empty, err := isDirEmpty(dir)
-		if err != nil {
-			if errors.Is(err, os.ErrNotExist) {
-				return nil
-			}
-			return err
-		}
-		if !empty {
+		if errors.Is(err, os.ErrNotExist) {
 			return nil
 		}
+		if err != nil || !empty {
+			return err
+		}
 
-		parentDir := path.Dir(dir)
 		if err := os.Remove(dir); err != nil {
 			return err
 		}
-		dir = parentDir
+		// move up to the parent dir
+		dir = path.Dir(dir)
 	}
-
 	return nil
 }
 
@@ -66,8 +62,8 @@ func (d *directory) createFile(path string) (*os.File, error) {
 	d.mu.Lock()
 	defer d.mu.Unlock()
 
-	if err := os.MkdirAll(filepath.Dir(path), os.ModePerm); err != nil {
-		return nil, fmt.Errorf("creating dir: %w", err)
+	if err := d.ensureParentDir(path); err != nil {
+		return nil, err
 	}
 
 	return os.Create(path)
@@ -78,9 +74,20 @@ func (d *directory) renameFile(src, dst string) error {
 	d.mu.Lock()
 	defer d.mu.Unlock()
 
-	if err := os.MkdirAll(filepath.Dir(dst), os.ModePerm); err != nil {
+	if err := d.ensureParentDir(dst); err != nil {
+		return err
+	}
+	return os.Rename(src, dst)
+}
+
+func (d *directory) ensureParentDir(path string) error {
+	parentDir := filepath.Dir(path)
+	if parentDir == d.ceilingDir {
+		return nil
+	}
+	err := os.MkdirAll(parentDir, os.ModePerm)
+	if err != nil {
 		return fmt.Errorf("creating dir: %w", err)
 	}
-
-	return os.Rename(src, dst)
+	return nil
 }
